chore: Stub refresh info provider
