+++ src/Uno.UI/Microsoft/UI/Xaml/Controls/PullToRefresh/RefreshContainer/Stub/StubRefreshInfoProvider.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.UI.Private.Controls;
+using Microsoft.UI.Xaml.Controls;
+using Windows.Foundation;
+using Windows.UI.Composition;
+
+namespace Uno.UI.Xaml.Controls;
+
+internal class StubRefreshInfoProvider : IRefreshInfoProvider
+{
+	public bool IsInteractingForRefresh => false;
+
+	public CompositionPropertySet CompositionProperties => null;
+
+	public string InteractionRatioCompositionProperty => null;
+
+	public double ExecutionRatio => 1;
+
+	public event TypedEventHandler<IRefreshInfoProvider, object> IsInteractingForRefreshChanged;
+
+	public event TypedEventHandler<IRefreshInfoProvider, RefreshInteractionRatioChangedEventArgs> InteractionRatioChanged;
+
+	public event TypedEventHandler<IRefreshInfoProvider, object> RefreshStarted;
+
+	public event TypedEventHandler<IRefreshInfoProvider, object> RefreshCompleted;
+
+	public void OnRefreshCompleted()
+	{
+	}
+
+	public void OnRefreshStarted()
+	{
+	}
+}
